Move api stuff to Shared folder
