+++ Shared/Extensions/Formatters.swift
@@ -1,0 +1,18 @@
+//
+//  Formatters.swift
+//  TreinRadar
+//
+//  Created by Wisse Hes on 06/08/2023.
+//
+
+import Foundation
+
+extension MeasurementFormatter {
+    static var myFormatter: MeasurementFormatter {
+        let formatter = self.init()
+        formatter.unitOptions = [.providedUnit, .naturalScale]
+        formatter.unitStyle = .long
+        formatter.numberFormatter.maximumFractionDigits = 0
+        return formatter
+    }
+}
