+++ Shared/Extensions/Journey+Computed.swift
