+++ Shared/Extensions/Location+Equatable.swift
