+++ Shared/Extensions/MessageStyle+String.swift
