+++ Shared/Extensions/Stop+Computed.swift
