+++ Shared/Extensions/String+Capitalization.swift
