+++ Shared/Structs/NS/CurrentVehicles.swift
