+++ Shared/Structs/NS/DeparturesResponse.swift
