+++ Shared/Structs/NS/JourneyGeojson.swift
