+++ Shared/Structs/NS/JourneyResponse.swift
