+++ Shared/Structs/NS/Shared.swift
