+++ Shared/Structs/NS/StationsResponse.swift
