+++ Shared/Structs/StationAnnotation.swift
