+++ Shared/Structs/StationWithDistance.swift
