+++ Shared/Structs/Stops+Geojson.swift
